backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/LLM Integration
openai==1.3.7
anthropic==0.7.7
langchain==0.0.340
jinja2==3.1.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
httpx==0.25.2
tenacity==8.2.3
pyyaml==6.0.1
pandas==2.1.3
numpy==1.26.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0